+++ Cycles_Calculator/src/PhysicsExtraCredit/CycleData.java
@@ -5,5 +5,5 @@ import java.util.ArrayList;
 public class CycleData {
 	ArrayList<CycleProcess> processData;
 	ArrayList<CycleNode> nodeData;
-	double moles;
+	float moles;
 }
