+++ Cycles_Calculator/src/PhysicsExtraCredit/IsoChor.java
@@ -5,36 +5,36 @@ public class IsoChor
 	
 	//create a method that calculates work,U,Q,and S 
 	//it must accept moles, temp, volume1, volume 2 as arguments
-	public static double isochorculationsWork()
+	public static float isochorculationsWork()
 	{
-		double work;
+		float work;
 	   //calculate the work which is nRT=ln(V2/V1)
 	    work = 0;
 	    
 	   return work;
 	}
 	
-	public static double isochorCalculationsQ(double Cv,double volume1,double Pfinal,double Pinitial)
+	public static float isochorCalculationsQ(float Cv,float volume1,float Pfinal,float Pinitial)
 	{
-		double Q;
+		float Q;
 		 Q= (Cv/Cycle.R)*(volume1 *(Pfinal - Pinitial));
 		 
 		 return Q;
 	}
 	
-	public static double isochorCalculationsS(double moles,double Cv, double tempFinal,double tempInitial)
+	public static float isochorCalculationsS(float moles,float Cv, float tempFinal,float tempInitial)
 	{
-		double S;
+		float S;
 		
 		//calculate the entropy for Free expansion
-		 S= (moles*Cv*(double)Math.log((tempFinal-tempInitial)));
+		 S= (moles*Cv*(float)Math.log((tempFinal-tempInitial)));
 		 
 		 return S;
 	}
 	
-	public static double isochorCalculationsU(double Cv,double volume1,double Pfinal,double Pinitial)
+	public static float isochorCalculationsU(float Cv,float volume1,float Pfinal,float Pinitial)
 	{
-		double U;
+		float U;
 		
 		//calculate the entropy for Free expansion
 		U= (Cv/Cycle.R)*(volume1 *(Pfinal - Pinitial));
@@ -43,9 +43,9 @@ public class IsoChor
 	}
 	
 
-	public static double isochorCalculationsTemp1(double moles,double volume1,double Pinitial)
+	public static float isochorCalculationsTemp1(float moles,float volume1,float Pinitial)
 	{
-		double temp1;
+		float temp1;
 		
 		//calculate the entropy for Free expansion
 		 temp1= ((Pinitial* volume1)/(moles*Cycle.R));
@@ -53,9 +53,9 @@ public class IsoChor
 		 return temp1;
 	}
 	
-	public static double isochorCalculationsTemp2(double moles,double volume1,double Pfinal)
+	public static float isochorCalculationsTemp2(float moles,float volume1,float Pfinal)
 	{
-		double temp2 ;
+		float temp2 ;
 		
 		//calculate the entropy for Free expansion
 		temp2= ((Pfinal*volume1)/(moles*Cycle.R));
