+++ Cycles_Calculator/src/PhysicsExtraCredit/SimpsonsRule.java
@@ -27,7 +27,7 @@ public class SimpsonsRule
    * Standard normal distribution density function.
    * Replace with any sufficiently smooth function.
    **********************************************************************/
-   public static double f(double x) 
+   public static float f(float x) 
    {
 	   //Math.exp(- x * x / 2) / Math.sqrt(2 * Math.PI);
       return  1/x;//Math.pow(x, 2); // Math.sqrt(2 * Math.PI);
@@ -38,25 +38,25 @@ public class SimpsonsRule
    * Integrate f from a to b using Simpson's rule.
    * Increase N for more precision.
    **********************************************************************/
-   public static double integrate(double a, double b) 
+   public static float integrate(float a, float b) 
    {
       int N = 10000;                    // precision parameter
-      double h = ((b - a) / (N - 1));     // step size
+      float h = ((b - a) / (N - 1));     // step size
  
       // 1/3 terms
-      double sum =  (1.0 / 3.0 * (f(a) + f(b)));
+      float sum =  (1.0f / 3.0f * (f(a) + f(b)));
 
       // 4/3 terms
       for (int i = 1; i < N - 1; i += 2) 
       {
-    	  double x = a + h * i;
+    	  float x = a + h * i;
          sum += 4.0 / 3.0 * f(x);
       }
 
       // 2/3 terms
       for (int i = 2; i < N - 1; i += 2) 
       {
-    	  double x = a + h * i;
+    	  float x = a + h * i;
          sum += 2.0 / 3.0 * f(x);
       }
